+++ Sources/StreamChatUI/CommonViews/SwiftUIViewRepresentable.swift
@@ -10,6 +10,7 @@ public protocol SwiftUIRepresentable: AnyObject {
     /// Type used for `content` property
     associatedtype ViewContent
     /// Content of a given view
+    @available(*, deprecated, message: "SwiftUIRepresentable is now deprecated, please refer to the SwiftUI SDK at https://github.com/GetStream/stream-chat-swiftui")
     var content: ViewContent { get set }
 }
 
@@ -18,6 +19,7 @@ public extension SwiftUIRepresentable where Self: UIView {
     /// Creates `SwiftUIViewRepresentable` instance wrapping the current type that can be used in your SwiftUI view
     /// - Parameters:
     ///     - content: Content of the view. Its value is automatically updated when it's changed
+    @available(*, deprecated, message: "SwiftUIRepresentable is now deprecated, please refer to the SwiftUI SDK at https://github.com/GetStream/stream-chat-swiftui")
     static func asView(_ content: ViewContent) -> SwiftUIViewRepresentable<Self> {
         SwiftUIViewRepresentable(
             view: self,
@@ -31,6 +33,7 @@ public extension SwiftUIRepresentable where Self: UIViewController {
     /// Creates `SwiftUIViewControllerRepresentable` instance wrapping the current type that can be used in your SwiftUI view
     /// - Parameters:
     ///     - content: Content of the view controller. Its value is automatically updated when it's changed
+    @available(*, deprecated, message: "SwiftUIRepresentable is now deprecated, please refer to the SwiftUI SDK at https://github.com/GetStream/stream-chat-swiftui")
     static func asView(_ content: ViewContent) -> SwiftUIViewControllerRepresentable<Self> {
         SwiftUIViewControllerRepresentable(
             viewController: self,
@@ -45,6 +48,7 @@ public struct SwiftUIViewRepresentable<View: UIView & SwiftUIRepresentable>: UIV
     private let view: View.Type
     private let content: View.ViewContent
     
+    @available(*, deprecated, message: "SwiftUIViewRepresentable is now deprecated, please refer to the SwiftUI SDK at https://github.com/GetStream/stream-chat-swiftui")
     init(
         view: View.Type,
         content: View.ViewContent
@@ -53,10 +57,12 @@ public struct SwiftUIViewRepresentable<View: UIView & SwiftUIRepresentable>: UIV
         self.content = content
     }
     
+    @available(*, deprecated, message: "SwiftUIViewRepresentable is now deprecated, please refer to the SwiftUI SDK at https://github.com/GetStream/stream-chat-swiftui")
     public func makeUIView(context: Context) -> View {
         view.init()
     }
     
+    @available(*, deprecated, message: "SwiftUIViewRepresentable is now deprecated, please refer to the SwiftUI SDK at https://github.com/GetStream/stream-chat-swiftui")
     public func updateUIView(_ uiView: View, context: Context) {
         uiView.content = content
     }
@@ -71,6 +77,7 @@ public struct SwiftUIViewControllerRepresentable<
     private let viewController: ViewController.Type
     private let content: ViewController.ViewContent
     
+    @available(*, deprecated, message: "SwiftUIViewControllerRepresentable is now deprecated, please refer to the SwiftUI SDK at https://github.com/GetStream/stream-chat-swiftui")
     init(
         viewController: ViewController.Type,
         content: ViewController.ViewContent
@@ -79,11 +86,13 @@ public struct SwiftUIViewControllerRepresentable<
         self.content = content
     }
     
+    @available(*, deprecated, message: "SwiftUIViewControllerRepresentable is now deprecated, please refer to the SwiftUI SDK at https://github.com/GetStream/stream-chat-swiftui")
     public func makeUIViewController(context: Context) -> ViewController {
         let controller = ViewController()
         controller.content = content
         return controller
     }
     
+    @available(*, deprecated, message: "SwiftUIViewControllerRepresentable is now deprecated, please refer to the SwiftUI SDK at https://github.com/GetStream/stream-chat-swiftui")
     public func updateUIViewController(_ uiViewController: ViewController, context: Context) {}
 }